requirements.txt
click==8.0.1
dataclasses==0.8
Flask==2.0.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.15.1
greenlet==1.1.0
importlib-metadata==4.6.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pkg_resources==0.0.0
python-dotenv==0.19.0
SQLAlchemy==1.3.23
typing-extensions==3.10.0.0
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.5.0
